+++ src/pages/test/index.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+
+interface ITestProps {
+};
+
+const Test = ({
+}: ITestProps): JSX.Element => {
+  return (
+    <div>
+      test pGe 2
+    </div>
+  );
+};
+
+export default Test
